Added LineageName component
